+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Renamed the item tag `#irregular_implements:fluid_tags` to `#irregular_implements:can_stand_on_fluids`, since that's what it's used for
 - The function to see if you can stand on a fluid now checks for the Curios `feet` slot
+- Made Spectre tree blocks translucent (#34)
 
 ### Added